added magnet declination correction for pose and absaccelerations, filtering sensors values,cleaned project and other small fixes

+++ app/src/main/java/com/app/carnavar/utils/filters/FiltersUtils.java
@@ -8,14 +8,14 @@ public class FiltersUtils {
     // dt - sec, timestamp - nanosec
 
     // low-pass filter, or exponentially weighted moving average, alpha=0.0..1.0
-    public static double[] lowPassFilteringArray(double[] newValuesVec, double[] prevValuesVec, double alpha) {
+    public static float[] lowPassFilteringArray(float[] newValuesVec, float[] prevValuesVec, float alpha) {
         for (int i = 0; i < newValuesVec.length; i++) {
             newValuesVec[i] = lowPassFiltering(newValuesVec[i], prevValuesVec[i], alpha);
         }
         return newValuesVec;
     }
 
-    public static double lowPassFiltering(double newValue, double prevValue, double alpha) {
+    public static float lowPassFiltering(float newValue, float prevValue, float alpha) {
         return alpha * newValue + (1 - alpha) * prevValue; // = prevValue + alpha * (newValue - prevValue)
     }
 }
